# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Fantasy Game API
  version: 1.0.0
servers:
- url: https://api.noobsupreme.com
paths:
  /campaigns:
    get:
      summary: Get a list of campaigns
      operationId: campaigns_get
      responses:
        "200":
          description: A list of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new campaign
      operationId: campaigns_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
        required: true
      responses:
        "201":
          description: Campaign created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /characters:
    get:
      summary: Get a list of characters
      operationId: characters_get
      responses:
        "200":
          description: A list of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new character
      operationId: characters_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
        required: true
      responses:
        "201":
          description: Character created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /dungeons:
    get:
      summary: List all dungeons
      operationId: dungeons_get
      responses:
        "200":
          description: A list of dungeons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dungeon'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new dungeon
      operationId: dungeons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dungeon'
        required: true
      responses:
        "201":
          description: New dungeon created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /dungeons/{dungeonId}:
    get:
      summary: Get a dungeon by ID
      operationId: dungeons_dungeon_id_get
      parameters:
      - name: dungeonId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Dungeon found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dungeon'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update a dungeon by ID
      operationId: dungeons_dungeon_id_put
      parameters:
      - name: dungeonId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dungeon'
        required: true
      responses:
        "200":
          description: Dungeon updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete a dungeon by ID
      operationId: dungeons_dungeon_id_delete
      parameters:
      - name: dungeonId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Dungeon deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /encounters:
    get:
      summary: Get a list of encounters
      operationId: encounters_get
      responses:
        "200":
          description: A list of encounters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Encounter'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new encounter
      operationId: encounters_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Encounter'
        required: true
      responses:
        "201":
          description: Encounter created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /encounters/{encounterId}:
    get:
      summary: Get an encounter by ID
      operationId: encounters_encounter_id_get
      parameters:
      - name: encounterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Encounter found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update an encounter by ID
      operationId: encounters_encounter_id_put
      parameters:
      - name: encounterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Encounter'
        required: true
      responses:
        "200":
          description: Encounter updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete an encounter by ID
      operationId: encounters_encounter_id_delete
      parameters:
      - name: encounterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Encounter deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /items:
    get:
      summary: Get a list of items
      operationId: items_get
      responses:
        "200":
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new item
      operationId: items_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "201":
          description: Item created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /items/{itemId}:
    get:
      summary: Get an item by ID
      operationId: items_item_id_get
      parameters:
      - name: itemId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update an item by ID
      operationId: items_item_id_put
      parameters:
      - name: itemId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "200":
          description: Item updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete an item by ID
      operationId: items_item_id_delete
      parameters:
      - name: itemId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Item deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /monsters:
    get:
      summary: Get a list of monsters
      operationId: monsters_get
      responses:
        "200":
          description: A list of monsters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monster'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new monster
      operationId: monsters_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Monster'
        required: true
      responses:
        "201":
          description: Monster created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /monsters/{monsterId}:
    get:
      summary: Get a monster by ID
      operationId: monsters_monster_id_get
      parameters:
      - name: monsterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Monster found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monster'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update a monster by ID
      operationId: monsters_monster_id_put
      parameters:
      - name: monsterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Monster'
        required: true
      responses:
        "200":
          description: Monster updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete a monster by ID
      operationId: monsters_monster_id_delete
      parameters:
      - name: monsterId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Monster deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /npcs:
    get:
      summary: Get a list of NPCs
      operationId: npcs_get
      responses:
        "200":
          description: A list of NPCs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NPC'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new NPC
      operationId: npcs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NPC'
        required: true
      responses:
        "201":
          description: NPC created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /npcs/{npcId}:
    get:
      summary: Get an NPC by ID
      operationId: npcs_npc_id_get
      parameters:
      - name: npcId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: NPC found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NPC'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update an NPC by ID
      operationId: npcs_npc_id_put
      parameters:
      - name: npcId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NPC'
        required: true
      responses:
        "200":
          description: NPC updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete an NPC by ID
      operationId: npcs_npc_id_delete
      parameters:
      - name: npcId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: NPC deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /quests:
    get:
      summary: Get a list of quests
      operationId: quests_get
      responses:
        "200":
          description: A list of quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quest'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new quest
      operationId: quests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quest'
        required: true
      responses:
        "201":
          description: Quest created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /quests/{questId}:
    get:
      summary: Get a quest by ID
      operationId: quests_quest_id_get
      parameters:
      - name: questId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Quest found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quest'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update a quest by ID
      operationId: quests_quest_id_put
      parameters:
      - name: questId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quest'
        required: true
      responses:
        "200":
          description: Quest updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete a quest by ID
      operationId: quests_quest_id_delete
      parameters:
      - name: questId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Quest deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /sessions:
    get:
      summary: Get a list of sessions
      operationId: sessions_get
      responses:
        "200":
          description: A list of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new session
      operationId: sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
        required: true
      responses:
        "201":
          description: Session created
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /sessions/{sessionId}:
    get:
      summary: Get a session by ID
      operationId: sessions_session_id_get
      parameters:
      - name: sessionId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Session found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Update a session by ID
      operationId: sessions_session_id_put
      parameters:
      - name: sessionId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
        required: true
      responses:
        "200":
          description: Session updated
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Delete a session by ID
      operationId: sessions_session_id_delete
      parameters:
      - name: sessionId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Session deleted
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /spells:
    get:
      summary: Get a list of spells
      operationId: spells_get
      responses:
        "200":
          description: A list of spells
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spell'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new spell
      operationId: spells_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Spell'
        required: true
      responses:
        "201":
          description: Spell created
      x-openapi-router-controller: swagger_server.controllers.default_controller
...
  /users:
    get:
      summary: Get a list of users
      operationId: users_get
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: Create a new user
      operationId: users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: User created
      x-openapi-router-controller: swagger_server.controllers.default_controller
...
components:
  schemas:
    Campaign:
      required:
      - campaignName
      - description
      - dungeonMaster
      - mainStory
      - players
      - startDate
      - world
      type: object
      properties:
        campaignName:
          type: string
        description:
          type: string
        dungeonMaster:
          type: string
        players:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        world:
          type: string
        mainStory:
          type: string
        sideQuests:
          type: array
          items:
            type: string
        majorEvents:
          type: array
          items:
            type: string
        notes:
          type: string
      example:
        world: world
        notes: notes
        endDate: 2000-01-23
        players:
        - players
        - players
        description: description
        sideQuests:
        - sideQuests
        - sideQuests
        mainStory: mainStory
        campaignName: campaignName
        majorEvents:
        - majorEvents
        - majorEvents
        dungeonMaster: dungeonMaster
        startDate: 2000-01-23
...